fixup sockaddr from recvfrom in case packet came over IPv4
The recvfrom returns the AF_INET in the sockaddr record
in case the packet comes via IPv4. All great, but sendto
does not like this on MacOS X and spews the "Invalid argument"
error.

Add the conversion to the IPv4-mapped IPv6 address and AF_INET6 for
the sockaddr_in6 - this makes the sendto work with IPv4 as well.